requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
defusedxml==0.7.1
et-xmlfile==1.1.0
fonttools==4.34.4
fpdf2==2.5.5
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.11
word2number==1.1
yapf==0.32.0
